requirements.txt
asgiref==3.3.0
astroid==2.4.2
colorama==0.4.3
Django==3.0.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
isort==5.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
